👷 build(repo): 由 turbo 统一接管任务编排
统一任务编排入口，消除 Taskfile 与 pnpm 脚本双轨调度：turbo 负责
缓存与并行调度，Taskfile 降级为薄委托层，保留需要 GITHUB_OUTPUT/
secrets/CLI 透传的任务（release/version/db/hooks/smoke）。

- lingc-cli 迁移至 apps/lingc-cli：uv workspace 成员 + turbo 节点
- turbo.json 注册 //#py:* / //#md:* / //#js:* / //#i18n / //#wheel-smoke
  根任务，显式排除缓存目录防止哈希污染
- release-lingc-cli.yml 发布路径同步；CI 经 task ci:* 自动继承
- docs：agent 三镜像与开发者文档同步目录约定与命令入口

+++ apps/lingc-cli/src/lingc_cli/handlers/run.py
@@ -167,7 +167,7 @@ async def _forward_output(
 async def _start_and_confirm(
     entry: list[str],
     cwd: Path,
-    timeout: int,
+    timeout: float,
 ) -> asyncio.subprocess.Process:
     """Spawn *entry*, forward output, and wait for a clean startup.
 
@@ -207,7 +207,7 @@ async def run(
     cmd: list[str],
     *,
     cwd: Path | None = None,
-    timeout: int = DEFAULT_STARTUP_TIMEOUT,
+    timeout: float = DEFAULT_STARTUP_TIMEOUT,
     reload: bool = False,
 ) -> int:
     """Safely start the bot and supervise it until it exits.
